ga alterrnativo 1

+++ classes/food.py
@@ -1,7 +1,7 @@
 class FoodNode:
     def __init__(self, protein, calories, fat, carbs, name):
-        self.protein: int = protein
-        self.calories: int = calories
-        self.carbs: int = carbs
-        self.fat: int = fat
+        self.protein: float = protein
+        self.calories: float = calories
+        self.carbs: float = carbs
+        self.fat: float = fat
         self.name: str = name
